refactor: relocate PagedResult<T> to Xcc.Persistance
Move the generic PagedResult<T> pagination DTO out of the Journal
feature's domain namespace and into Anela.Heblo.Xcc.Persistance,
alongside IRepository<TEntity, TKey>. Removes the cross-module
dependency where Marketing's domain interface imported
Anela.Heblo.Domain.Features.Journal solely to reach PagedResult.

Pure relocation — type shape, property names, serialization, and
runtime behavior are unchanged. No new abstractions, no shim, no
project references added.
