+++ backend/src/Anela.Heblo.Domain/Features/Journal/IJournalRepository.cs
@@ -22,12 +22,4 @@ namespace Anela.Heblo.Domain.Features.Journal
             IEnumerable<string> productCodes,
             CancellationToken cancellationToken = default);
     }
-
-    public class PagedResult<T>
-    {
-        public List<T> Items { get; set; } = new();
-        public int TotalCount { get; set; }
-        public int PageNumber { get; set; }
-        public int PageSize { get; set; }
-    }
 }
