+++ backend/src/Anela.Heblo.Persistence/Catalog/Journal/JournalRepository.cs
@@ -1,5 +1,6 @@
 using Anela.Heblo.Domain.Features.Journal;
 using Anela.Heblo.Persistence.Repositories;
+using Anela.Heblo.Xcc.Persistance;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
