+++ backend/src/Anela.Heblo.Xcc/Persistance/PagedResult.cs
@@ -1,0 +1,9 @@
+namespace Anela.Heblo.Xcc.Persistance;
+
+public class PagedResult<T>
+{
+    public List<T> Items { get; set; } = new();
+    public int TotalCount { get; set; }
+    public int PageNumber { get; set; }
+    public int PageSize { get; set; }
+}
